finish ai advice and handle new user navigation
