+++ components/title.tsx
@@ -1,17 +1,29 @@
 import React from "react";
-import { Text, StyleSheet, View, TouchableOpacity } from "react-native";
+import { Text, StyleSheet, View, TouchableOpacity, Image } from "react-native";
 import { ChevronLeft } from "lucide-react-native";
 
 interface TitleProps {
   text: string;
   backBottom: () => void;
+  logoutFunction?: boolean;
 }
 
-const Title: React.FC<TitleProps> = ({ text, backBottom }) => {
+const Title: React.FC<TitleProps> = ({
+  text,
+  backBottom,
+  logoutFunction = false,
+}) => {
   return (
     <View style={styles.container}>
       <TouchableOpacity onPress={backBottom} style={styles.leftButton}>
-        <ChevronLeft size={24} color="#000000" />
+        {logoutFunction ? (
+          <Image
+            source={require("@/assets/images/logoutIcon.png")}
+            style={styles.logoutIcon}
+          />
+        ) : (
+          <ChevronLeft size={24} color="#000000" />
+        )}
       </TouchableOpacity>
       <Text style={styles.title}>{text}</Text>
       <View style={styles.placeholder} />
@@ -38,6 +50,10 @@ const styles = StyleSheet.create({
   },
   placeholder: {
     flex: 1,
+  },
+  logoutIcon: {
+    width: 32,
+    height: 32,
   },
 });
